pyproject.toml
[tool.ruff]
output-format = "full"
lint.select = ["ALL"]
line-length = 100
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    ".eggs",
    "*.egg",
    "dist",
    "tests/fixtures/**",
    "tests/**/snapshots/**",
    "alembic",
    "airich",
]
lint.ignore = [
    "D100",
    "B008",
    "D104",
    "Q000",
    "S101",
    "PT016",
    "ANN101",
    "ANN102",
    "N805",
    "PLC0414",
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"
ignore-decorators = ["typing.overload"]

[tool.ruff.lint.mccabe]
max-complexity = 11

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.Query"]

[tool.ruff.lint.extend-per-file-ignores]
"typings/*" = ["ANN401"]
"__init__.py" = ["F401", "F405"]
"*/migrations/versions/*" = ["D103"]
"src/app/main.py" = ["E402"]
"src/cli/commands/*" = ["E402", "ANN201"]
"src/cli/__init__.py" = ["E402", "ANN201", "FBT001", "FBT002", "UP007"]
"manage.py" = ["E402"]
"tests/*" = ["D101", "D102", "D103"]


[tool.black]
line-length = 100
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3


[tool.coverage]
[tool.coverage.run]
source = ["verbose_http_exceptions"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "enum.Enum",
    "(Protocol):",
    "(typing.Protocol):",
    "pragma: no cover",
    "pragma: no coverage",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "@overload",
]

[tool.pytest]
testpath = "tests"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.django.test_app.core.settings"
# python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]

[tool.bandit]
exclude = ['tests']

[tool.bandit.assert_used]
skips = ['*_test.py', '*/test_*.py']

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.6.1",
    "vulture>=2.11",
    "pytest>=8.3.2",
    "black>=24.8.0",
    "bandit>=1.7.9",
    "coverage>=7.6.1",
    "pytest-cov>=5.0.0",
    "isort>=5.13.2",
    "pyright>=1.1.376",
    "freezegun>=1.5.1",
    "mimesis>=17.0.0",
    "ipython>=8.26.0",
    "pytest-asyncio>=0.23.8",
    "httpx>=0.27.0",
    "pytest-mock>=3.14.0",
    "uvicorn>=0.30.6",
    "djangorestframework-types>=0.8.0",
    "pytest-django>=4.8.0",
]


[project]
name = "verbose_http_exceptions"
version = "3.1.0"
description = "Verbose HTTP exceptions package for web frameworks."
authors = [{ name = "Dmitriy Lunev", email = "dima.lunev14@gmail.com" }]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
dependencies = ["python-dev-utils>=7.0.1"]

[project.optional-dependencies]
fastapi = ["fastapi"]
litestar = ["litestar"]
django = ["django", "djangorestframework"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
